updated tests for select-multi and moved the function into validators for easier testing and because it makes some sense there.
