+++ ckanext/cfpb_extrafields/validators.py
@@ -10,15 +10,15 @@ def dedupe_unordered(items):
 def contains_bad_chars(str):
     ''' contains \"{},'''
     test = False
-    for c in ['\"','{','}',',']:
+    for c in ['>','<']:
         test = test or (c in str)
     return test
 def input_value_validator(value):
     # __Other option is the option that triggers a js event for user-specified option creation
     if "__Other" in value :
         Invalid("'Other, please specify' is not a valid option")
-#    if contains_bad_chars(value):
-#        Invalid('multi-select fields cannot contain commas, {}, or double-quotes')
+    if contains_bad_chars(value):
+        Invalid('multi-select fields cannot contain <, > or quote characters')
     # assume that duplicates are mistakes continue quietly
     if not isinstance(value, basestring):
         value = dedupe_unordered(value)
@@ -83,3 +83,45 @@ def reasonable_date_validator(value):
         except ValueError:
             Invalid("Please ensure date is in yyyy-mm-dd format.")
     return value
+
+
+# select multis are contained in unicode strings that look like:
+# u'{"blah blah","blah asdf",asdf}' ; u'{asdf,asdf}' ; u'asdf' (see also tests.py)
+def clean_select_multi0(a):
+    ''' parses the results of an html form select-multi '''
+    # this solution doesn't accomodate commas in select-multi fields
+    # and validation preventing users from entering commas is not straightforward.
+    return a.replace('{', "").replace("}", "").replace("\"","").split(",")
+def clean_select_multi(s):
+    ''' parses the results of an html form select-multi '''
+    # This solution allows commas, but is unpythonic
+    if not s:
+        return []
+    s = s.lstrip('{').rstrip('}')
+    clean = []
+    left = 0
+    right = 1
+    while right < len(s):
+        if s[left]=="\"":
+            if right+1<len(s) and s[right]=="\"" and s[right+1]==",":
+                # reached the end of the quoted patch; append and skip
+                clean.append(s[left+1:right])
+                left = right+2
+                right = right+3
+            else:
+                right = right+1
+        else:
+            if s[right]==",":
+                # hit a comma: append and skip the comma
+                clean.append(s[left:right])
+                left = right+1
+                right = right+2
+            else:
+                right = right+1
+    if s[left]=="\"":
+        # reached the end with a closing quote (don't save the quotes)
+        clean.append(s[left+1:-1])
+    else:
+        clean.append(s[left:])
+    return clean
+
